reorginized app.py by moving db into db.py and the routes into their specific files.
