+++ meal_max/meal_max/db.py
@@ -1,3 +1,52 @@
+from flask import Flask
 from flask_sqlalchemy import SQLAlchemy
+from werkzeug.security import generate_password_hash, check_password_hash
+import os
 
 db = SQLAlchemy()
+
+def create_app():
+    app = Flask(__name__)
+    app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///calorie_tracker.db'
+    app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
+    db.init_app(app)
+    return app
+
+class User(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    username = db.Column(db.String(50), unique=True, nullable=False)
+    calorie_goal = db.Column(db.Integer, nullable=False)
+    starting_weight = db.Column(db.Integer, nullable=False)
+    salt = db.Column(db.String(128), nullable=False)
+    password_hash = db.Column(db.String(128), nullable=False)
+
+    def __init__(self, username, password, calorie_goal, starting_weight):
+        self.username = username
+        self.salt = self.generate_salt()
+        self.password_hash = self.generate_password_hash(password)
+        self.calorie_goal = calorie_goal
+        self.starting_weight = starting_weight
+
+    def generate_salt(self):
+        """Generate a random salt"""
+        return os.urandom(16).hex()
+
+    def generate_password_hash(self, password):
+        """Generate password hash using salt"""
+        return generate_password_hash(password + self.salt)
+
+    def check_password(self, password):
+        """Check if the provided password matches the stored hash"""
+        return check_password_hash(self.password_hash, password + self.salt)
+    
+class WeightLog(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    user_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
+    date = db.Column(db.Date, nullable=False)
+    weight = db.Column(db.Float, nullable=False)
+
+class CalorieIntake(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    user_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
+    date = db.Column(db.Date, nullable=False)
+    calories = db.Column(db.Integer, nullable=False)
